Nicer abstraction for fetching triples
Although we use a class to coordinate the flow for fetching triples,
there is little use for that class except for maintaining state during
the fetching process.  A helper method has been introduced to take care
of the creation of an instance and calling the necessary methods on that
instance.