Split INSERT DATA and DELETE DATA queries into chunks
These queries could contain a lot of triples.  In some cases the number
goes over 1000.  This may or may not be a problem for the triplestore.
It greatly depends on the configuration.  Updating this service to
support sending triples in chunks will somewhat slow down the
processing, but it will increase stability.