Max amount of properties to be queried in a single call
We can send out batches of requests to the triplestore.  However, as the
amount of triples rises, the calls become heavier on the triplestore.
In order to alleviate this, we batch the amount of properties to fetch
to a fairly high maximum.  That should help in exploding queries in
Virtuoso.